break dog creation form into own component
